requirements.txt
black
flake8
pytest
requests
selenium
tox
python-dotenv
webdriver-manager
pytest-xdist
